api/model_service/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.6.0
numpy>=1.21.0
Pillow>=8.3.0
scikit-learn>=1.0.0
flask>=2.0.0
flask-cors>=3.0.0
requests>=2.26.0
werkzeug>=2.0.0
python-dotenv>=1.1
gunicorn>=20.0.0
pillow-heif